requirements.txt
python-telegram-bot[socks]
psycopg2-binary
python-dotenv
reportlab
google-api-python-client
google-auth-oauthlib
google-auth
google-generativeai